requirements.txt
Django>=6.0
Pillow>=10.0.0
django-environ>=0.13.0
gunicorn>=25.0
psycopg~=3.3.3
python_slugify~=8.0.4